rm pthreads from refcounting

+++ gc/mark_and_compact/test/bitset.c
@@ -612,7 +612,7 @@ register char *s;
 	static set a;
 	register unsigned *p, *endp;
 
-	set_new(a, strlen(s));
+	set_new(a, (int)strlen(s));
 	p = a.setword;
 	endp = &(a.setword[a.n]);
 	do {
